+++ download.sh
@@ -3,7 +3,7 @@
 # huggingface-cli login --token $HUGGINGFACE_TOKEN
 # export CURL_CA_BUNDLE=""
 # export REQUESTS_CA_BUNDLE=""
-HF_ENDPOINT=http://hf-mirror.com huggingface-cli download --repo-type dataset --resume-download BUAADreamer/clevr_count_70k --local-dir /storage/openpsi/data/clevr_count_70k 
+HF_ENDPOINT=http://hf-mirror.com huggingface-cli download --repo-type dataset --resume-download hiyouga/geometry3k --local-dir /storage/openpsi/data/geometry3k 
 # huggingface-cli download htlou/obelics_obelics_100k_tokenized_2048 --local-dir ./obelics_obelics_100k_tokenized_2048 --repo-type dataset
 # huggingface-cli download htlou/obelics_obelics_10k_tokenized_2048 --local-dir ./obelics_obelics_10k_tokenized_2048 --repo-type dataset
 # huggingface-cli download htlou/obelics_obelics_100k --local-dir ./obelics_obelics_100k --repo-type dataset
