+++ examples/arealite/dataset/gsm8k.py
@@ -1,4 +1,4 @@
-from datasets import load_dataset
+from datasets import load_dataset, Dataset
 from datasets.distributed import split_dataset_by_node
 
 def get_gsm8k_sft_dataset(path, split, tokenizer, rank, world_size):
@@ -17,4 +17,12 @@ def get_gsm8k_sft_dataset(path, split, tokenizer, rank, world_size):
     dataset = dataset.map(process).remove_columns(["question", "answer"])
     return dataset
 
+def get_gsm8k_rl_dataset(path,split, rank, world_size):
+    dataset = load_dataset(path=path, name="main", split=split)
+    dataset = split_dataset_by_node(dataset, rank=rank, world_size=world_size)
+    def process(sample):
+        messages = [{"role": "user", "content": sample["question"]}]
+        return {"messages": messages}
 
+    dataset = dataset.map(process).remove_columns(["question"])
+    return dataset
